börjar funka

+++ 04/lab_04/Matrix.h
@@ -498,8 +498,8 @@ std::istream & operator>>(std::istream & is, Matrix<T> & m) {
     int rows, cols;
     is >> rows >> cols; // read the number of rows and columns from the input stream
     m = Matrix<T>(rows, cols); //  set the matrix to the correct size
-    for (int i = 0; i < rows; i++) { // loop through the rows
-        for (int j = 0; j < cols; j++) { // loop through the columns
+    for (size_t i = 0; i < rows; i++) { // loop through the rows
+        for (size_t j = 0; j < cols; j++) { // loop through the columns
             is >> m(i, j); // read the value from the input stream
         }
     }
@@ -510,8 +510,8 @@ template<typename T>
 std::ostream & operator<<(std::ostream & os, const Matrix<T> & m) {
     // implement operator << to print a matrix to an output stream, that is like when we use cout << m;
     os << m.rows() << " " << m.cols() << endl; // print the number of rows and columns to the output stream
-    for (int i = 0; i < m.rows(); i++) { // loop through the rows
-        for (int j = 0; j < m.cols(); j++) { // loop through the columns
+    for (size_t i = 0; i < m.rows(); i++) { // loop through the rows
+        for (size_t j = 0; j < m.cols(); j++) { // loop through the columns
             os << m(i, j) << " "; // print the value to the output stream
         }
         os << endl; // print a new line
